pyproject.toml
[tool.poetry]
name = "japanese-lipnet"
version = "0.1.0"
description = "Japanese Lip Reading implementation with Conformer architecture"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "japanese_lipnet"}]

[tool.poetry.dependencies]
python = "~3.10"
torch = "1.13.1"
torchvision = "0.14.1"
numpy = "1.26.0"
opencv-python = "^4.10.0.84"
editdistance = "^0.8.1"
tensorboardx = "^2.6.2.2"
pandas = "^2.1.1"
tqdm = "^4.66.1"
jaconv = "^0.3.4"
scikit-learn = "^1.3.1"
matplotlib = "^3.8.0"
mediapipe = "^0.10.18"
tensorboard = "^2.18.0"
pyyaml = "^6.0.2"
python-levenshtein = "^0.26.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
pylint = "^3.0.1"
pytest = "^7.4.2"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "C0330",  # bad-continuation
    "C0326",  # bad-whitespace
]

[tool.pylint.format]
max-line-length = 100